Check rsc stonith is in maintenance by either of 2 keywords
Newer version of crmsh will report that a stonith resource is in
maintenance by the `maintenance` keyword, while older version use
`unmanaged` instead.

This was addressed in commit a88d1c545e7c1eb9297e4e974bde50fc3e8a525a,
but that failed to address the fact that the `check_cluster_conf_ssh`
function checks that the output matches all provided keywords. This has
led to many false failures, as output will never match both `unmanaged`
and `maintenance` at the same time.

This commit provides a fix by adding a third parameter to
`check_cluster_conf_ssh` so the function is able to match all provided
keywords, or just one of them. The extra parameter informs the function
whether it should do an "or" (set to false, i.e., "and", by default) and
return True when any of the keywords is matched, or do an "and" and
match all as it was doing before.

This commit also improves the log messages to reference both `unmanaged`
and `maintenance`.